# Speed up eslint with --cache