--- a/.gitignore
+++ b/.gitignore
@@ -51,3 +51,6 @@ RUNNING_PID
 # Docker-compose / docker sbt
 /docker-compose.yml
 /?
+
+# eslint cache
+/.eslintcache